数据结构支持long类型

+++ samples/unity_sample/json_sample/Assets/Scripts/Tables/PlayerLevel.cs
@@ -19,6 +19,6 @@ namespace Gable
         /// <summary>
         /// 血量
         /// </summary>
-        public int hp;
+        public long hp;
     }
 }
